rating

+++ Backend/idealTrip/src/main/java/com/idealtrip/idealTrip/service/ReviewService.java
@@ -26,13 +26,13 @@ public class ReviewService {
         return reviews.findAll();
     }
 
-    public Double getTotalRatingForDestination(Destination destination) {
+    public int getTotalRatingForDestination(Destination destination) {
         List<Review> reviewList = reviews.findByDestination(destination);
         int totalRating = 0;
         for (Review review : reviewList) {
             totalRating += review.getRatingReview();
         } 
-        return (double) totalRating / reviewList.size(); 
+        return (int) totalRating / reviewList.size(); 
     }
 
     public Page<Review> findReviewByUser(User user, Pageable pageable) {
